+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # New
 - `harpy view envs`: print versions
-
+- CLI validations for `--hpc`/`-H` to check if the plugins necessary for the configuration are installed
 
 # Fixes
 - `harpy view envs`: simpler logic and print diagnostic text if empty